third_party/WebKit/LayoutTests/editing/selection/modify_extend/extend_with_unfocused_selection.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => {
  assert_own_property(window, 'eventSender', 'This test requires eventSender');
  assert_selection(
  [
    '<textarea>|text</textarea>',
    '<a href="http://www.example.com/">link</a>'
  ].join(''),
  selection => {
    const link = selection.document.querySelector('a');
    link.focus();
    eventSender.keyDown('ArrowRight', ['shiftKey']);
    assert_equals(selection.document.activeElement, link);
  },
  [
    '|<textarea>text</textarea>',
    '<a href="http://www.example.com/">link</a>'
  ].join(''));
}, 'Press Shift + right arrow key with unfocused selection in text control');

test(() => {
  assert_own_property(window, 'eventSender', 'This test requires eventSender');
  assert_selection(
  [
    '<div contenteditable>|text</div>',
    '<a href="http://www.example.com/">link</a>'
  ].join(''),
  selection => {
    const link = selection.document.querySelector('a');
    link.focus();
    eventSender.keyDown('ArrowRight', ['shiftKey']);
    assert_equals(selection.document.activeElement, link);
  },
  [
    '<div contenteditable>|text</div>',
    '<a href="http://www.example.com/">link</a>'
  ].join(''));
}, 'Press Shift + right arrow key with unfocused selection in contenteditable div');
</script>
